+++ syncrclone/main.py
@@ -137,14 +137,16 @@ class SyncRClone:
         )
         log("")
         log(f"A >>> B {self.sumA}")
-        self.rclone.transfer("A2B", self.transA2B)
+        
+        self.rclone.transfer("A2B", *self.split_transfer_lists_matching_size('A2B') )
 
+        self.split_transfer_lists_matching_size('B2A')
         self.sumB = utils.file_summary(
             [self.currB.query_one(Path=f) for f in self.transB2A]
         )
         log("")
         log(f"A <<< B {self.sumB}")
-        self.rclone.transfer("B2A", self.transB2A)
+        self.rclone.transfer("B2A", *self.split_transfer_lists_matching_size('B2A') )
 
         # Update lists if needed
         log("")
@@ -440,9 +442,9 @@ class SyncRClone:
             sA, sB = fileA["Size"], fileB["Size"]
 
             txtA = utils.unix2iso(mA) if mA else "<< not found >>"
-            txtA += " ({:0.2g} {:s})".format(*utils.bytes2human(sA, short=True))
+            txtA += f" ({sA:d} bytes)"
             txtB = utils.unix2iso(mB) if mB else "<< not found >>"
-            txtB += " ({:0.2g} {:s})".format(*utils.bytes2human(sB, short=True))
+            txtB += f" ({sB:d} bytes)"
 
             if config.conflict_mode not in {"newer", "older", "newer_tag"}:
                 mA, mB = sA, sB  # Reset m(AB) to s(AB)
@@ -666,6 +668,36 @@ class SyncRClone:
             return True  # Only got here size is equal
 
         return abs(file1["mtime"] - file2["mtime"]) <= config.dt
+
+    def split_transfer_lists_matching_size(self,mode):
+        """
+        Split transfers into whether they match size or not. See documentation
+        of rclone.transfer for explanation
+        """
+        if mode == 'A2B':
+            trans = self.transA2B
+            src = self.currA
+            dst = self.currB
+        elif mode == 'B2A':
+            trans = self.transB2A
+            src = self.currB
+            dst = self.currA
+        else:
+            raise ValueError('bad mode')
+        
+        matched_size = []
+        diff_size = []
+        
+        for file in trans:
+            fsrc = src.query_one(Path=file)
+            fdst = dst.query_one(Path=file)
+            
+            if not fdst or fsrc['Size'] != fdst['Size']:
+                diff_size.append(file)
+            else:
+                matched_size.append(file)
+        
+        return matched_size,diff_size
 
     def avoid_relist(self):
         # actions: 'new','del','tag','backup','trans','moves'